Added 2 new Projects
The code interpreter project is able to check the code the user puts in and the assistend can better analyse and help with the code because it knows the errors directly (only works with pythen code currently)

The secont project introduces the main features needed for voice interaction which are text to speach and speach to text. tts was easie but wispers mice to text is more tricky and still needs to be optemized. however it works already.

because wisper gets complicated there are two wisper files a simple one that listens for 7 seconds and a more complex one wich listents continously.

Most of the code was written with help of gpt4 Turbo. I also used gpt4o from time to time if turbo seemed to get stuck with an error